+++ weaver/processes/wps_process_base.py
@@ -105,9 +105,9 @@ class WpsProcessInterface(abc.ABC):
         The function is expected to monitor the process and update the status.
         Retrieve the expected outputs and store them in the ``out_dir``.
 
-        :param workflow_inputs: `CWL` job dict
-        :param out_dir: directory where the outputs must be written
-        :param expected_outputs: expected value outputs as `{'id': 'value'}`
+        :param workflow_inputs: :term:`CWL` job dictionary.
+        :param out_dir: directory where the outputs must be written.
+        :param expected_outputs: expected value outputs as ``{"id": "value"}``.
         """
         self.update_status("Preparing process for remote execution.",
                            RemoteJobProgress.PREPARE, Status.RUNNING)